refactor: extract shared Admin/Owner components into generic shared components
