+++ src/app/pages/OwnerPanel.tsx
@@ -31,347 +31,10 @@ import { getPrimaryProductImage, getProductImages } from "../utils/productImages
 import QRCode from "react-qr-code";
 import { getCategoryColors } from "../utils/categoryStyle";
 
-// ── SHARED COMPONENTS ──────────────────────────────────────────────────────
-
-function Modal({
-  title,
-  onClose,
-  children,
-}: {
-  title: string;
-  onClose: () => void;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onClose}
-      />
-      <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-lg max-h-[90vh] overflow-y-auto">
-        <div className="flex items-center justify-between p-5 border-b border-gray-100 sticky top-0 bg-white rounded-t-2xl z-10">
-          <h3
-            className="text-gray-900"
-            style={{
-              fontFamily: "Poppins, sans-serif",
-              fontWeight: 700,
-              fontSize: "1.05rem",
-            }}
-          >
-            {title}
-          </h3>
-          <button
-            onClick={onClose}
-            className="p-2 rounded-xl hover:bg-gray-100 text-gray-500 transition-colors"
-          >
-            <X className="w-5 h-5" />
-          </button>
-        </div>
-        <div className="p-5">{children}</div>
-      </div>
-    </div>
-  );
-}
-
-function Field({
-  label,
-  children,
-}: {
-  label: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <div>
-      <label
-        className="block text-gray-700 mb-1"
-        style={{
-          fontFamily: "Nunito, sans-serif",
-          fontWeight: 700,
-          fontSize: "0.85rem",
-        }}
-      >
-        {label}
-      </label>
-      {children}
-    </div>
-  );
-}
-
-const inputCls =
-  "w-full px-3 py-2.5 rounded-xl border border-gray-200 focus:border-purple-500 focus:ring-2 focus:ring-purple-100 outline-none transition-all bg-gray-50 focus:bg-white text-gray-800";
-
-function ConfirmDialog({
-  message,
-  onConfirm,
-  onCancel,
-}: {
-  message: string;
-  onConfirm: () => void;
-  onCancel: () => void;
-}) {
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-      <div
-        className="absolute inset-0 bg-black/50 backdrop-blur-sm"
-        onClick={onCancel}
-      />
-      <div className="relative bg-white rounded-2xl shadow-2xl w-full max-w-sm p-6 text-center">
-        <div className="w-12 h-12 rounded-full bg-red-100 flex items-center justify-center mx-auto mb-4">
-          <Trash2 className="w-6 h-6 text-red-600" />
-        </div>
-        <p
-          className="text-gray-800 mb-6"
-          style={{ fontFamily: "Nunito, sans-serif", fontWeight: 600 }}
-        >
-          {message}
-        </p>
-        <div className="flex gap-3">
-          <button
-            onClick={onConfirm}
-            className="flex-1 py-2.5 rounded-xl bg-red-600 text-white font-bold hover:bg-red-700 transition-colors"
-            style={{ fontFamily: "Nunito, sans-serif" }}
-          >
-            Confirmar
-          </button>
-          <button
-            onClick={onCancel}
-            className="flex-1 py-2.5 rounded-xl border-2 border-gray-200 text-gray-600 font-bold hover:bg-gray-50 transition-colors"
-            style={{ fontFamily: "Nunito, sans-serif" }}
-          >
-            Cancelar
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ── PRODUCT FORM ───────────────────────────────────────────────────────────
-function ProductForm({
-  initial,
-  onSave,
-  onClose,
-}: {
-  initial?: Partial<Product>;
-  onSave: (data: Partial<Product>) => Promise<void> | void;
-  onClose: () => void;
-}) {
-  const initialPriceMode = resolvePriceMode(initial || {});
-  const initialImages = getProductImages(initial || {});
-  const [form, setForm] = useState({
-    name: initial?.name || "",
-    description: initial?.description || "",
-    priceMode: initialPriceMode,
-    price:
-      initialPriceMode === "single" ? initial?.price?.toString() || "" : "",
-    priceMin:
-      initialPriceMode === "range"
-        ? initial?.priceMin?.toString() || initial?.price?.toString() || ""
-        : "",
-    priceMax:
-      initialPriceMode === "range"
-        ? initial?.priceMax?.toString() || initial?.price?.toString() || ""
-        : "",
-    image: getPrimaryProductImage(initial || {}),
-    images: initialImages,
-  });
-  const set = (k: string, v: string) => setForm((f) => ({ ...f, [k]: v }));
-  const setImages = (images: string[]) => {
-    setForm((f) => ({ ...f, images, image: images[0] || "" }));
-  };
-  const [productNameError, setProductNameError] = useState("");
-  const [productPriceError, setProductPriceError] = useState("");
-
-  const handleSave = async () => {
-    setProductNameError("");
-    setProductPriceError("");
-    if (!form.name) {
-      setProductNameError("Campo obrigatório");
-      throw new Error("Validation Error");
-    }
-    if (form.priceMode === "single") {
-      if (!form.price) {
-        setProductPriceError("Campo obrigatório");
-        throw new Error("Validation Error");
-      }
-      const singlePrice = parseFloat(form.price.replace(",", "."));
-      if (Number.isNaN(singlePrice) || singlePrice < 0) {
-        setProductPriceError("Informe um preço válido");
-        throw new Error("Validation Error");
-      }
-      await onSave({
-        name: form.name,
-        description: form.description,
-        image: form.image,
-        images: form.images,
-        priceMode: "single",
-        price: singlePrice,
-      });
-      return;
-    }
-    if (form.priceMode === "range") {
-      if (!form.priceMin || !form.priceMax) {
-        setProductPriceError("Preencha mínimo e máximo");
-        throw new Error("Validation Error");
-      }
-      const min = parseFloat(form.priceMin.replace(",", "."));
-      const max = parseFloat(form.priceMax.replace(",", "."));
-      if (Number.isNaN(min) || Number.isNaN(max) || min < 0 || max < 0) {
-        setProductPriceError("Informe uma faixa válida");
-        throw new Error("Validation Error");
-      }
-      if (min > max) {
-        setProductPriceError("O mínimo não pode ser maior que o máximo");
-        throw new Error("Validation Error");
-      }
-      await onSave({
-        name: form.name,
-        description: form.description,
-        image: form.image,
-        images: form.images,
-        priceMode: "range",
-        price: min,
-        priceMin: min,
-        priceMax: max,
-      });
-      return;
-    }
-    await onSave({
-      name: form.name,
-      description: form.description,
-      image: form.image,
-      images: form.images,
-      priceMode: "hidden",
-      price: 0,
-    });
-  };
-
-  return (
-    <div className="space-y-4">
-      <Field label="Nome do produto *">
-        <input
-          className={inputCls}
-          value={form.name}
-          onChange={(e) => set("name", e.target.value)}
-          onBlur={() => {
-            if (!form.name) setProductNameError("Campo obrigatório");
-            else setProductNameError("");
-          }}
-          placeholder="Nome"
-          style={{ fontFamily: "Nunito, sans-serif" }}
-        />
-        {productNameError && (
-          <p
-            className="text-red-600 text-sm mt-1"
-            style={{ fontFamily: "Nunito, sans-serif", fontWeight: 700 }}
-          >
-            {productNameError}
-          </p>
-        )}
-      </Field>
-      <Field label="Descrição">
-        <textarea
-          className={inputCls}
-          value={form.description}
-          onChange={(e) => set("description", e.target.value)}
-          rows={3}
-          placeholder="Descreva o produto"
-          style={{ fontFamily: "Nunito, sans-serif" }}
-        />
-      </Field>
-      <Field label="Exibição de preço *">
-        <select
-          className={inputCls}
-          value={form.priceMode}
-          onChange={(e) => {
-            set("priceMode", e.target.value);
-            setProductPriceError("");
-          }}
-          style={{ fontFamily: "Nunito, sans-serif" }}
-        >
-          <option value="single">Preço único</option>
-          <option value="range">Faixa de preço</option>
-          <option value="hidden">Não mostrar preço</option>
-        </select>
-      </Field>
-      {form.priceMode === "single" && (
-        <Field label="Preço (R$) *">
-          <input
-            className={inputCls}
-            type="text"
-            inputMode="decimal"
-            value={form.price}
-            onChange={(e) => set("price", e.target.value)}
-            onBlur={() => {
-              if (!form.price) setProductPriceError("Campo obrigatório");
-              else setProductPriceError("");
-            }}
-            placeholder="0,00"
-            style={{ fontFamily: "Nunito, sans-serif" }}
-          />
-        </Field>
-      )}
-      {form.priceMode === "range" && (
-        <div className="grid grid-cols-2 gap-3">
-          <Field label="Preço mínimo (R$) *">
-            <input
-              className={inputCls}
-              type="text"
-              inputMode="decimal"
-              value={form.priceMin}
-              onChange={(e) => set("priceMin", e.target.value)}
-              placeholder="0,00"
-              style={{ fontFamily: "Nunito, sans-serif" }}
-            />
-          </Field>
-          <Field label="Preço máximo (R$) *">
-            <input
-              className={inputCls}
-              type="text"
-              inputMode="decimal"
-              value={form.priceMax}
-              onChange={(e) => set("priceMax", e.target.value)}
-              placeholder="0,00"
-              style={{ fontFamily: "Nunito, sans-serif" }}
-            />
-          </Field>
-        </div>
-      )}
-      {productPriceError && (
-        <p
-          className="text-red-600 text-sm mt-1"
-          style={{ fontFamily: "Nunito, sans-serif", fontWeight: 700 }}
-        >
-          {productPriceError}
-        </p>
-      )}
-      <Field label="Fotos do produto">
-        <ProductImageGalleryField
-          value={form.images}
-          onChange={setImages}
-        />
-      </Field>
-      <div className="flex gap-3 pt-2">
-        <SubmitButton
-          onClick={handleSave}
-          className="flex-1 flex items-center justify-center gap-2 py-2.5 rounded-xl text-white font-bold shadow-md transition-all hover:scale-[1.01]"
-          style={{
-            background: "linear-gradient(135deg, #7C3AED, #EA580C)",
-            fontFamily: "Nunito, sans-serif",
-          }}
-          idleText="Salvar produto"
-        />
-        <button
-          onClick={onClose}
-          className="flex items-center gap-2 px-4 py-2.5 rounded-xl font-bold text-sm border-2 border-gray-200 text-gray-600 hover:bg-gray-50 transition-all"
-          style={{ fontFamily: "Nunito, sans-serif" }}
-        >
-          Cancelar
-        </button>
-      </div>
-    </div>
-  );
-}
+import { Modal } from "../components/shared/Modal";
+import { Field, inputCls } from "../components/shared/Field";
+import { ConfirmDialog } from "../components/shared/ConfirmDialog";
+import { ProductForm } from "../components/shared/ProductForm";
 
 // ── ENTERPRISE EDIT FORM ───────────────────────────────────────────────────
 function EnterpriseEditForm({
